Adjusted signature of libxsmm_coprime2 (uint->size_t).

+++ include/libxsmm_math.h
@@ -70,7 +70,7 @@ LIBXSMM_API size_t libxsmm_lcm(size_t a, size_t b);
 LIBXSMM_API int libxsmm_primes_u32(unsigned int num, unsigned int num_factors_n32[]);
 
 /** Calculate co-prime number <= n/2 (except: libxsmm_coprime2(0|1) == 0). */
-LIBXSMM_API size_t libxsmm_coprime2(unsigned int n);
+LIBXSMM_API size_t libxsmm_coprime2(size_t n);
 
 /**
  * Minimizes the waste, if "a" can only be processed in multiples of "b".
